LRCLyrics support 🎉
Custom useragent, custom method to generate image avg color (for karaoke mode background, inferior to the colors spotify gives us). we manually run lrclyric fetching and avg color fetching if spotify gives us a 404 (either user is on a Free account or lyric is missing. either way LRCLyrics is worth a shot). We also add a new init to SpotifyColorData.